refactor(tests): mover testes de thumbnail_worker para subdiretório
Move test_thumbnail_{app,extractor,handler,storage}.py para
unit/workers/thumbnail_worker/ e extrai fixture _mock_pg para
conftest.py local.

Refs: #136
