+++ tests/unit/workers/thumbnail_worker/test_extractor.py
