+++ tests/unit/workers/thumbnail_worker/test_handler.py
