+++ tests/unit/workers/thumbnail_worker/test_storage.py
